requirements.txt
gym==0.21.0
matplotlib==3.5.1
numpy==1.22.1
scipy==1.7.3
setuptools==59.5.0
opencv-python==4.5.5.62
pynput==1.7.6
Pillow==9.0.1
pycocotools==2.0.4
